Refactored backoff policies module
