--- src/backoff-policies/_module-exports.ts
@@ -1,6 +1,0 @@
-export * from "@/backoff-policies/constant-backoff/_module.js";
-export * from "@/backoff-policies/exponential-backoff/_module.js";
-export * from "@/backoff-policies/linear-backoff/_module.js";
-export * from "@/backoff-policies/polynomial-backoff/_module.js";
-export * from "@/backoff-policies/_shared.js";
-export * from "@/backoff-policies/types.js";
