--- src/backoff-policies/constant-backoff/_module.ts
@@ -1,1 +1,0 @@
-export * from "@/backoff-policies/constant-backoff/constant-backoff.js";
