--- src/backoff-policies/exponential-backoff/_module.ts
@@ -1,1 +1,0 @@
-export * from "@/backoff-policies/exponential-backoff/exponential-backoff.js";
