+++ src/backoff-policies/implementations/linear-backoff/linear-backoff.test.ts
@@ -1,39 +1,36 @@
 import { describe, expect, test } from "vitest";
 
 import {
-    polynomialBackoff,
-    resolvePolynomialBackoffSettings,
-} from "@/backoff-policies/polynomial-backoff/polynomial-backoff.js";
+    linearBackoff,
+    resolveLinearBackoffSettings,
+} from "@/backoff-policies/implementations/linear-backoff/linear-backoff.js";
 import { TO_MILLISECONDS } from "@/time-span/contracts/_module.js";
 import { TimeSpan } from "@/time-span/implementations/_module.js";
 import { callInvokable } from "@/utilities/_module.js";
 
-describe("function: resolvePolynomialBackoffSettings", () => {
+describe("function: resolveLinearBackoffSettings", () => {
     test("Should use default values when no settings are provided", () => {
-        const resolved = resolvePolynomialBackoffSettings({});
+        const resolved = resolveLinearBackoffSettings({});
 
         expect(resolved.maxDelay[TO_MILLISECONDS]()).toBe(60_000);
         expect(resolved.minDelay[TO_MILLISECONDS]()).toBe(500);
-        expect(resolved.degree).toBe(2);
         expect(resolved.jitter).toBe(0.5);
     });
 
     test("Should use provided values", () => {
-        const resolved = resolvePolynomialBackoffSettings({
+        const resolved = resolveLinearBackoffSettings({
             maxDelay: TimeSpan.fromSeconds(30),
             minDelay: TimeSpan.fromMilliseconds(100),
-            degree: 3,
             jitter: 0.2,
         });
 
         expect(resolved.maxDelay[TO_MILLISECONDS]()).toBe(30_000);
         expect(resolved.minDelay[TO_MILLISECONDS]()).toBe(100);
-        expect(resolved.degree).toBe(3);
         expect(resolved.jitter).toBe(0.2);
     });
 
     test("Should allow null jitter", () => {
-        const resolved = resolvePolynomialBackoffSettings({
+        const resolved = resolveLinearBackoffSettings({
             jitter: null,
         });
 
@@ -42,7 +39,7 @@ describe("function: resolvePolynomialBackoffSettings", () => {
 
     test("Should throw TypeError when minDelay is zero", () => {
         expect(() =>
-            resolvePolynomialBackoffSettings({
+            resolveLinearBackoffSettings({
                 minDelay: TimeSpan.fromMilliseconds(0),
             }),
         ).toThrow("'minDelay' must be positive");
@@ -50,7 +47,7 @@ describe("function: resolvePolynomialBackoffSettings", () => {
 
     test("Should throw TypeError when minDelay is negative", () => {
         expect(() =>
-            resolvePolynomialBackoffSettings({
+            resolveLinearBackoffSettings({
                 minDelay: TimeSpan.fromMilliseconds(-100),
             }),
         ).toThrow("'minDelay' must be positive");
@@ -58,7 +55,7 @@ describe("function: resolvePolynomialBackoffSettings", () => {
 
     test("Should throw TypeError when minDelay is NaN", () => {
         expect(() =>
-            resolvePolynomialBackoffSettings({
+            resolveLinearBackoffSettings({
                 minDelay: TimeSpan.fromMilliseconds(NaN),
             }),
         ).toThrow("'minDelay' must be positive");
@@ -66,7 +63,7 @@ describe("function: resolvePolynomialBackoffSettings", () => {
 
     test("Should throw TypeError when maxDelay is NaN", () => {
         expect(() =>
-            resolvePolynomialBackoffSettings({
+            resolveLinearBackoffSettings({
                 maxDelay: TimeSpan.fromMilliseconds(NaN),
             }),
         ).toThrow("'maxDelay' must be greater than or equal to 'minDelay'");
@@ -74,40 +71,16 @@ describe("function: resolvePolynomialBackoffSettings", () => {
 
     test("Should throw TypeError when maxDelay is less than minDelay", () => {
         expect(() =>
-            resolvePolynomialBackoffSettings({
+            resolveLinearBackoffSettings({
                 maxDelay: TimeSpan.fromMilliseconds(100),
                 minDelay: TimeSpan.fromMilliseconds(200),
             }),
         ).toThrow("'maxDelay' must be greater than or equal to 'minDelay'");
     });
 
-    test("Should throw TypeError when degree is zero", () => {
-        expect(() =>
-            resolvePolynomialBackoffSettings({
-                degree: 0,
-            }),
-        ).toThrow("'degree' must be positive");
-    });
-
-    test("Should throw TypeError when degree is negative", () => {
-        expect(() =>
-            resolvePolynomialBackoffSettings({
-                degree: -1,
-            }),
-        ).toThrow("'degree' must be positive");
-    });
-
-    test("Should throw TypeError when degree is NaN", () => {
-        expect(() =>
-            resolvePolynomialBackoffSettings({
-                degree: NaN,
-            }),
-        ).toThrow("'degree' must be positive");
-    });
-
     test("Should throw TypeError when jitter is negative", () => {
         expect(() =>
-            resolvePolynomialBackoffSettings({
+            resolveLinearBackoffSettings({
                 jitter: -0.1,
             }),
         ).toThrow("'jitter' must be between 0 and 1 or null");
@@ -115,7 +88,7 @@ describe("function: resolvePolynomialBackoffSettings", () => {
 
     test("Should throw TypeError when jitter is greater than 1", () => {
         expect(() =>
-            resolvePolynomialBackoffSettings({
+            resolveLinearBackoffSettings({
                 jitter: 1.1,
             }),
         ).toThrow("'jitter' must be between 0 and 1 or null");
@@ -123,80 +96,59 @@ describe("function: resolvePolynomialBackoffSettings", () => {
 
     test("Should throw TypeError when jitter is NaN", () => {
         expect(() =>
-            resolvePolynomialBackoffSettings({
+            resolveLinearBackoffSettings({
                 jitter: NaN,
             }),
         ).toThrow("'jitter' must be between 0 and 1 or null");
     });
 });
 
-describe("function: polynomialBackoff", () => {
-    test("Should grow polynomially with each attempt", () => {
-        const backoff = polynomialBackoff({
+describe("function: linearBackoff", () => {
+    test("Should grow linearly with each attempt", () => {
+        const backoff = linearBackoff({
             minDelay: TimeSpan.fromMilliseconds(500),
             maxDelay: TimeSpan.fromSeconds(60),
-            degree: 2,
             jitter: null,
         });
 
-        // 500 * 1^2 = 500
+        // 500 * 1 = 500
         expect(callInvokable(backoff, 1, undefined)[TO_MILLISECONDS]()).toBe(
             500,
         );
-        // 500 * 2^2 = 2000
+        // 500 * 2 = 1000
         expect(callInvokable(backoff, 2, undefined)[TO_MILLISECONDS]()).toBe(
-            2000,
+            1000,
         );
-        // 500 * 3^2 = 4500
+        // 500 * 3 = 1500
         expect(callInvokable(backoff, 3, undefined)[TO_MILLISECONDS]()).toBe(
-            4500,
+            1500,
         );
     });
 
     test("Should cap at maxDelay", () => {
-        const backoff = polynomialBackoff({
+        const backoff = linearBackoff({
             minDelay: TimeSpan.fromMilliseconds(500),
-            maxDelay: TimeSpan.fromMilliseconds(3000),
-            degree: 2,
+            maxDelay: TimeSpan.fromMilliseconds(1200),
             jitter: null,
         });
 
-        // 500 * 3^2 = 4500, capped at 3000
+        // 500 * 3 = 1500, capped at 1200
         expect(callInvokable(backoff, 3, undefined)[TO_MILLISECONDS]()).toBe(
-            3000,
-        );
-    });
-
-    test("Should use custom degree", () => {
-        const backoff = polynomialBackoff({
-            minDelay: TimeSpan.fromMilliseconds(100),
-            maxDelay: TimeSpan.fromSeconds(60),
-            degree: 3,
-            jitter: null,
-        });
-
-        // 100 * 2^3 = 800
-        expect(callInvokable(backoff, 2, undefined)[TO_MILLISECONDS]()).toBe(
-            800,
-        );
-        // 100 * 3^3 = 2700
-        expect(callInvokable(backoff, 3, undefined)[TO_MILLISECONDS]()).toBe(
-            2700,
+            1200,
         );
     });
 
     test("Should apply jitter to the delay", () => {
-        const backoff = polynomialBackoff({
+        const backoff = linearBackoff({
             minDelay: TimeSpan.fromMilliseconds(500),
             maxDelay: TimeSpan.fromSeconds(60),
-            degree: 2,
             jitter: 0.5,
             _mathRandom: () => 0.5,
         });
 
-        // attempt=2: 500 * 2^2 = 2000, jitter: (1 - 0.5 * 0.5) * 2000 = 1500
+        // attempt=2: 500 * 2 = 1000, jitter: (1 - 0.5 * 0.5) * 1000 = 750
         expect(callInvokable(backoff, 2, undefined)[TO_MILLISECONDS]()).toBe(
-            1500,
+            750,
         );
     });
 });
