--- src/backoff-policies/polynomial-backoff/_module.ts
@@ -1,1 +1,0 @@
-export * from "@/backoff-policies/polynomial-backoff/polynomial-backoff.js";
